refactor: move ChipGroup to flat level
