--- packages/blade/src/components/Chip/ChipGroup/index.ts
@@ -1,1 +1,0 @@
-export * from './ChipGroup';
