+++ packages/blade/src/components/Chip/index.ts
@@ -1,2 +1,4 @@
-export * from './Chip';
-export * from './ChipGroup';
+export { Chip } from './Chip';
+export { ChipGroup } from './ChipGroup';
+export type { ChipProps } from './Chip';
+export type { ChipGroupProps } from './ChipGroup';
